Reject null path parameters for C api calls
It's easier to check this inside the library instead of at every caller,
which would be necessary if the argument is a nullable field extracted
from another call.

This problem was very visible for ArmAPI_v1GetInstance() in the example,
because the library incorrectly sets the id to null in case of error. I
need to deal with that next.