Don't set returned C structures with error info
After a call to the web api, C api functions parse the returned json to
populate a structure that is returned to the caller. The problem is that
the json may not actually match this structure: if the remote call
failed, we will just have an error description. The result is that even
failed calls may report success, but return a struct with all null
fields.

To avoid this, check the status code before parsing the json. It would
be nice to still provide the error details to the user somehow, I'll
look into it later.